pyproject.toml
[tool.poetry]
name = "sequant"
version = "0.1.0"
description = ""
authors = ["nserov"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
biopython = "^1.83"
absl-py = "^2.1.0"
rdkit = "^2023.9.4"
pandas = "^2.1.4"
scikit-learn = "^1.3.2"
tensorflow-io-gcs-filesystem = "^0.31.0"
tensorflow = "~2.8"
rdkit-pypi = "^2022.9.5"
peptides = "^0.3.2"
tqdm = "^4.66.1"
python-dotenv = "^1.0.1"
fastapi = "^0.112.0"
numpy = "1.26.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
